# api docs perf improvement