Trying to root out perf issues with the api routes. The following is based on averages of 10 trials

**Setup**

- run `NO_CACHE=1 npm run dev`
- Disable Cache in chrome dev tools
- Go to http://localhost:3000/dev/api/react-router

**Results**

| Loader           | Before (ms) | After (ms) |
| ---------------- | ----------- | ---------- |
| `api-layout.tsx` | 735.80      | 380.338    |